requirements.txt
streamlit
opencv-python
numpy
pillow
pytesseract
pdf2image
pandas
matplotlib
requests
python-dotenv
google-auth
google-generativeai
streamlit_option_menu
spacy
plotly
#python -m spacy download en_core_web_sm